Refactor: UserCommand 메소드 퍼사드패턴으로 리팩토링
