+++ src/main/java/com/boardgo/domain/user/service/UserCommandUseCase.java
@@ -1,19 +1,21 @@
 package com.boardgo.domain.user.service;
 
 import com.boardgo.domain.user.controller.request.SignupRequest;
-import com.boardgo.domain.user.controller.request.SocialSignupRequest;
 import com.boardgo.domain.user.controller.request.UserPersonalInfoUpdateRequest;
+import com.boardgo.domain.user.entity.UserInfoEntity;
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface UserCommandUseCase {
-    Long signup(SignupRequest signupRequest);
+    Long save(SignupRequest signupRequest);
 
-    Long socialSignup(SocialSignupRequest socialSignupRequest, Long userId);
+    void validateNickNameAndPrTag(String nickName, List<String> prTags);
+
+    UserInfoEntity getUserInfoEntity(Long userId);
+
+    boolean existsUser(String nickname);
 
     void updatePersonalInfo(Long userId, UserPersonalInfoUpdateRequest updateRequest);
 
     void updateProfileImage(Long userId, MultipartFile profileImage);
-
-    void updatePrTags(List<String> changedPrTag, Long userId);
 }
