+++ src/main/java/com/boardgo/domain/user/service/UserPrTagCommandUseCase.java
@@ -1,0 +1,9 @@
+package com.boardgo.domain.user.service;
+
+import java.util.List;
+
+public interface UserPrTagCommandUseCase {
+    void bulkInsertPrTags(List<String> prTags, Long userInfoId);
+
+    void updatePrTags(List<String> changedPrTag, Long userId);
+}
